Update error and warning messages
Error and warning messages are now automatically styled with "Warning: "/"ERROR: " prefixes. Both now allow the user to press any key to continue so that they can acknowledge the error. Error messages now also have the option to add a message to contact the creator in the case of an error.